requirements.txt
azure-storage-blob
python-multipart  
uvicorn  
fastapi
opencv-python